moved mkldnn_func_kernel to .cc
