+++ onnxruntime/core/providers/mkldnn/subgraph/mkldnn_batchnorm.h
@@ -4,7 +4,6 @@
 #pragma once
 #include "core/framework/op_kernel.h"
 #include "core/providers/mkldnn/mkldnn_fwd.h"
-#include "core/providers/cpu/nn/autopad_type.h"
 #include "core/providers/mkldnn/mkldnn_execution_provider.h"
 #include "core/providers/mkldnn/subgraph/mkldnn_kernel.h"
 #include "core/providers/mkldnn/memcpy_s.h"
