+++ onnxruntime/core/providers/mkldnn/subgraph/mkldnn_lrn.h
@@ -6,7 +6,6 @@
 #include "core/util/math_cpuonly.h"
 #include "core/framework/op_kernel.h"
 #include "core/providers/mkldnn/mkldnn_fwd.h"
-#include "core/providers/cpu/nn/autopad_type.h"
 #include "core/providers/mkldnn/mkldnn_execution_provider.h"
 #include "core/providers/mkldnn/subgraph/mkldnn_kernel.h"
