+++ onnxruntime/core/providers/mkldnn/subgraph/mkldnn_sum.h
@@ -5,7 +5,6 @@
 #include "core/framework/op_kernel.h"
 #include "core/providers/mkldnn/mkldnn_fwd.h"
 #include "core/providers/mkldnn/mkldnn_common.h"
-#include "core/providers/cpu/nn/autopad_type.h"
 #include "core/providers/mkldnn/subgraph/mkldnn_kernel.h"
 #include "core/util/math.h"
